\name{QM16}
\alias{QM16}
\title{QTL MAS - Workshop 16 dataset}
\description{

This is part of the QTL-MAS 16 Workshop dataset consisting of three phenotypes
and 10000 markers for 3000 individuals. 

}

\usage{
  data(QM16)
}

\format{
\code{Y} is a matrix of 3000 observations and 3 columns, each column representing one phenotype.
\code{M} is a matrix of 3000 observations and 10000 columns (marker-covariates).
}

\source{
http://qtl-mas-2012.kassiopeagroup.com/en/program.php
}

\references{
http://qtl-mas-2012.kassiopeagroup.com/en/program.php
}
\keyword{datasets}
